_backup_1759473442/requirements.txt
Flask==3.0.3
python-magic==0.4.27
gunicorn==22.0.0
PyJWT==2.8.0
coincurve==17.0.0
cryptography==41.0.3
